Use common make variable names
This commit replaces the GCC_FLAGS variable with the names used in the
implicit make rules: CFLAGS, CPPFLAGS, and CXXFLAGS. The first reason
for such change is that those names are better known in the world of
make users and are should _make_ the scripts easier to understand for
newcomers. The more important reason is that they actually separate the
C compiler flags, the C++ compiler flags, and the C/C++ preprocessor
flags from each other. It should then be easier to construct a rule for
building a C‑only (or C++‑only) project.

See node 10.3 of the make manual, “Variables Used by Implicit Rules”.

This commit effectively partially reverts the commit
92ca8f233c573f8e68b6c357b8ea8c8b3b49a0bf.